Feat: Connect Sign-In API and set token at cookies
