+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^2.5.0",
+    "axios": "^1.7.9",
     "next": "14.2.5",
     "react": "^18",
     "react-dom": "^18",
